openbmc-subtree: fix reset bugs
Fix several bugs in the reset function that were fixed in the update
function previously:

Fix several exceptions where a non-existent helper method 'git()' was
being called.  These are left over from a previous refactor and the
reset method was not updated or tested after the refactor.  We need to
use the git method in args, which is appropriately scoped to the target
specified by the -t option, or invoke sh.git directly.

Normalize the path to the config file.  This avoids a bug when the
config file is in the same directory as the working directory for the
script.

Fix generating commit messages for resets - these did not work at all.